Adjusted document layout parameters in PDF generator
The use of 'SimpleDocTemplate' in the functions for photo index building in 'hff_system__exp_shipwrecksheet_pdf.py' has been modified. The pagesize parameters were changed from A4 standard to a specific cm-based size and 'showBoundary' property was set to 0. The 'canvasmaker' argument was also updated. This adjustment is made to improve the layout and appearance of the generated PDFs.